+++ src/kernel/asm.c
@@ -1,13 +1,13 @@
 #include "asm.h"
 
-void write_port(uint8 port, uint8 value)
+void write_port(uint16 port, uint16 value)
 {
     __asm__ volatile("outb %%al,%%dx" ::"a"(value), "d"(port));
 }
 
-uint8 read_port(uint8 port)
+uint16 read_port(uint16 port)
 {
-    uint8 _v;
+    uint16 _v;
     __asm__ volatile("inb %%dx,%%al"
                      : "=a"(_v)
                      : "d"(port));
